feat: support decimal calories and serving size input for ingredients
  Allow users to enter calories per serving (e.g. "165 cal per 100g")
  instead of manually computing per-unit values. The app divides
  calories by serving size before storing. This enables ingredients
  like broccoli (0.34 cal/g) that were previously impossible to add.

  - Changed calories_per_unit from INT to NUMERIC in schema
  - Added serving size field to ingredient form with live preview
  - Fixed Decimal × float TypeError in meal calorie calculations
  - Rounded calorie totals to 2 decimal places to prevent noise
  - Added backend tests for decimal calories and validation

+++ infra/sql/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE ingredients (
   id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
   user_id UUID REFERENCES users(id) ON DELETE CASCADE,
   name VARCHAR(255) NOT NULL,
-  calories_per_unit INT NOT NULL CHECK (calories_per_unit >= 0),
+  calories_per_unit NUMERIC NOT NULL CHECK (calories_per_unit >= 0),
   unit VARCHAR(50) NOT NULL
 );
